+++ cloudssifier/train.py
@@ -8,8 +8,6 @@ from absl import app
 from absl import flags
 from absl import logging
 
-import os
-
 import numpy as np
 
 import tensorflow as tf
@@ -17,17 +15,18 @@ from tensorflow.keras.models import Sequential
 from tensorflow.keras.layers import Activation
 from tensorflow.keras.layers import Dense
 from tensorflow.keras.optimizers import SGD
-import tensorflow.keras.backend as K
+
 from tensorflow.python.client import device_lib
 
 FLAGS = flags.FLAGS
 
-default_data_path = '../../data/mnist.npz'
+default_data_path = 'data/mnist.npz'
 
 flags.DEFINE_integer('epochs', 10, 'number of training epochs.')
 flags.DEFINE_integer('batch_size', 64, 'training batch size.')
 flags.DEFINE_string('data_path', default_data_path, 'path to mnist npz file.')
-flags.DEFINE_string('job-dir',  '',  'Catch job directory passed by cloud.')
+flags.DEFINE_string('job-dir', '', 'Catch job directory passed by cloud.')
+
 
 def get_mnist(data_path):
   """
@@ -39,15 +38,15 @@ def get_mnist(data_path):
     data - Tuple of tuples, consisting of formatted train and test data.
   """
   logging.info('Attempting to get data from %s' % data_path)
-  with tf.gfile.GFile(data_path, mode='rb') as f:
+  with tf.io.gfile.GFile(data_path, mode='rb') as f:
     data = np.load(f)
 
   x_train, y_train_cold = data['x_train'], data['y_train']
   x_test, y_test_cold = data['x_test'], data['y_test']
 
-  x_train = x_train.reshape(-1, 28*28)/255.0
+  x_train = x_train.reshape(-1, 28 * 28) / 255.0
   x_train = x_train.astype(np.float32)
-  x_test = x_test.reshape(-1, 28*28)/255.0
+  x_test = x_test.reshape(-1, 28 * 28) / 255.0
   x_test = x_test.astype(np.float32)
 
   l_train = len(y_train_cold)
@@ -55,12 +54,13 @@ def get_mnist(data_path):
 
   y_train = np.zeros((l_train, 10), dtype=np.float32)
   y_test = np.zeros((l_test, 10), dtype=np.float32)
-  y_train[np.arange(l_train),y_train_cold] = 1
-  y_test[np.arange(l_test),y_test_cold] = 1
+  y_train[np.arange(l_train), y_train_cold] = 1
+  y_test[np.arange(l_test), y_test_cold] = 1
 
   data = ((x_train, y_train), (x_test, y_test))
 
   return data
+
 
 def get_model():
   """
@@ -72,7 +72,7 @@ def get_model():
 
   # Model
   model = Sequential()
-  model.add(Dense(256, input_shape=(28*28,)))
+  model.add(Dense(256, input_shape=(28 * 28,)))
   model.add(Activation('relu'))
   model.add(Dense(256))
   model.add(Activation('relu'))
@@ -83,8 +83,9 @@ def get_model():
   optimizer = SGD(lr=0.1)
 
   # Compile
-  model.compile(optimizer=optimizer, loss='categorical_crossentropy',
-              metrics=['accuracy'])
+  model.compile(optimizer=optimizer,
+                loss='categorical_crossentropy',
+                metrics=['accuracy'])
 
   return model
 
@@ -108,24 +109,25 @@ def train_and_validate():
 
   logging.info('Validating model.')
 
-  sess = K.get_session()
-
-  train_preds = np.argmax(model(x_train).eval(session=sess), axis=1)
+  train_preds = np.argmax(model(x_train).numpy(), axis=1)
   train_labels = np.argmax(y_train, axis=1)
   train_acc = np.mean(train_preds == train_labels)
 
-  test_preds = np.argmax(model(x_test).eval(session=sess), axis=1)
+  test_preds = np.argmax(model(x_test).numpy(), axis=1)
   test_labels = np.argmax(y_test, axis=1)
   test_acc = np.mean(test_preds == test_labels)
 
   print('Train acc: %f, Test acc: %f' % (train_acc, test_acc))
 
+
 def main(argv):
   logging.info('Running main.')
+
   if len(argv) > 1:
     raise app.UsageError('Too many command-line arguments.')
-  else:
-    train_and_validate()
+
+  train_and_validate()
+
 
 if __name__ == '__main__':
   app.run(main)
