+++ setup.cfg
@@ -1,0 +1,3 @@
+[yapf]
+based_on_style = google
+indent_width = 2
